requirements.txt
alembic==1.13.3
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
DateTime==5.5
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.1
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.10
PyJWT==2.9.0
pytest==8.3.3
pytz==2024.2
setuptools==75.1.0
SQLAlchemy==2.0.35
SQLAlchemy-Utils==0.41.2
typing_extensions==4.12.2
Werkzeug==3.0.4
zope.interface==7.1.0
